Removed INetworkModule\NetworkModule and Utilities Cleanup
- INetworkModule\NetworkModule removed and relevant code merged into Game
- Utilities project removed and relevant code merged into respective areas of Game
- Auth\Upload Clients no longer check for partial packets as Auth Emu will never send any
