+++ Game/Maps/Entities/TerrainPropInfo.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.IO;
 using Navislamia.Game.Maps.Enums;
-using Navislamia.Utilities;
 
 namespace Navislamia.Game.Maps.Entities;
