--- Game/Network/INetworkModule.cs
@@ -1,7 +1,0 @@
-﻿namespace Navislamia.Game.Network
-{
-    public interface INetworkModule
-    {
-        bool Start();
-    }
-}
