+++ Game/Network/Packets/Enums/AuthPackets.cs
