+++ Game/Network/Packets/PacketExtensions.cs
@@ -1,30 +1,10 @@
-﻿using Navislamia.Network.Packets;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
+﻿using System;
 using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading.Tasks;
-using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Navislamia.Game.Network.Packets;
 
 public static class PacketExtensions
 {
-    public static Header PeekHeader(this byte[] data)
-    {
-        if (data.Length < 7)
-            throw new Exception("Not enough data to peek header!");
-
-        var _header = new Header();
-        _header.Length = BitConverter.ToUInt32(data, 0);
-        _header.ID = BitConverter.ToUInt16(data, 4);
-        _header.Checksum = data[6];
-
-        return _header;
-    }
-
     /// <summary>
     /// Calculate the checksum of the packet
     /// Ported from Glandu2 CLI Packet Serializer
