+++ Navislamia.sln
@@ -6,10 +6,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DevConsole", "DevConsole\De
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Configuration", "Configuration\Configuration.csproj", "{200B7054-AD61-44E3-AC3D-02EE58AAE86F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Utilities", "Utilities\Utilities.csproj", "{B5A58ABD-5B0A-4FAE-BBC7-94C19ED19956}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Objects", "Objects\Objects.csproj", "{F6595EE8-2E3A-4484-9F9F-EEC7546E3100}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Game", "Game\Game.csproj", "{1C1B438A-9255-45CC-85BB-730BAFFE4D7E}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Tests\Tests.csproj", "{8BD11344-88EB-401A-AD74-D30F8A0465F0}"
@@ -42,22 +38,6 @@ Global
 		{200B7054-AD61-44E3-AC3D-02EE58AAE86F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{200B7054-AD61-44E3-AC3D-02EE58AAE86F}.Release|x64.ActiveCfg = Release|Any CPU
 		{200B7054-AD61-44E3-AC3D-02EE58AAE86F}.Release|x64.Build.0 = Release|Any CPU
-		{B5A58ABD-5B0A-4FAE-BBC7-94C19ED19956}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B5A58ABD-5B0A-4FAE-BBC7-94C19ED19956}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B5A58ABD-5B0A-4FAE-BBC7-94C19ED19956}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{B5A58ABD-5B0A-4FAE-BBC7-94C19ED19956}.Debug|x64.Build.0 = Debug|Any CPU
-		{B5A58ABD-5B0A-4FAE-BBC7-94C19ED19956}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B5A58ABD-5B0A-4FAE-BBC7-94C19ED19956}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B5A58ABD-5B0A-4FAE-BBC7-94C19ED19956}.Release|x64.ActiveCfg = Release|Any CPU
-		{B5A58ABD-5B0A-4FAE-BBC7-94C19ED19956}.Release|x64.Build.0 = Release|Any CPU
-		{F6595EE8-2E3A-4484-9F9F-EEC7546E3100}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F6595EE8-2E3A-4484-9F9F-EEC7546E3100}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F6595EE8-2E3A-4484-9F9F-EEC7546E3100}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{F6595EE8-2E3A-4484-9F9F-EEC7546E3100}.Debug|x64.Build.0 = Debug|Any CPU
-		{F6595EE8-2E3A-4484-9F9F-EEC7546E3100}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F6595EE8-2E3A-4484-9F9F-EEC7546E3100}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F6595EE8-2E3A-4484-9F9F-EEC7546E3100}.Release|x64.ActiveCfg = Release|Any CPU
-		{F6595EE8-2E3A-4484-9F9F-EEC7546E3100}.Release|x64.Build.0 = Release|Any CPU
 		{1C1B438A-9255-45CC-85BB-730BAFFE4D7E}.Debug|Any CPU.ActiveCfg = Debug|x64
 		{1C1B438A-9255-45CC-85BB-730BAFFE4D7E}.Debug|Any CPU.Build.0 = Debug|x64
 		{1C1B438A-9255-45CC-85BB-730BAFFE4D7E}.Debug|x64.ActiveCfg = Debug|x64
