+++ Tests/Tests.csproj
@@ -22,8 +22,6 @@
       <ProjectReference Include="..\Configuration\Configuration.csproj" />
       <ProjectReference Include="..\DevConsole\DevConsole.csproj" />
       <ProjectReference Include="..\Game\Game.csproj" />
-      <ProjectReference Include="..\Objects\Objects.csproj" />
-      <ProjectReference Include="..\Utilities\Utilities.csproj" />
     </ItemGroup>
 
 </Project>
